# gateway-service/api/docs/swagger.yaml
definitions:
...
  handlers.CreateCustomerRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  handlers.CreateCustomerResponse:
    properties:
      customer_id:
        type: string
      message:
        type: string
    required:
    - customer_id
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.ListAccountResponse:
    properties:
      accounts: {}
      message:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
      totalPages:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/customer:
    get:
      consumes:
      - application/json
      description: Get customer list
      parameters:
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      - default: 50
        description: Number of customers per page
        in: query
        name: pagesize
        type: integer
      - description: Bearer token for authorization, include 'Bearer ' followed by
          access_token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Get Customer List
      tags:
      - Customer
    post:
      consumes:
      - application/json
      description: Create customer - Bearer token required
      parameters:
      - description: Bearer token for authorization, include 'Bearer ' followed by
          access_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Customer details
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateCustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create Customer
      tags:
      - Customer
  /api/v1/customer/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a customer by customer id
      parameters:
      - description: Bearer token for authorization, include 'Bearer ' followed by
          access_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: CustomerID of the customer
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Delete Customer
      tags:
      - Customer
  /api/v1/customer/{id}/account:
    get:
      consumes:
      - application/json
      description: Get account list of customers
      parameters:
      - description: CustomerID of the customer
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      - default: 50
        description: Number of accounts per page
        in: query
        name: pagesize
        type: integer
      - description: Bearer token for authorization, include 'Bearer ' followed by
          access_token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved account list
          schema:
            $ref: '#/definitions/handlers.ListAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get Account List of customer
      tags:
      - Customer
...
swagger: "2.0"